Defined access modifiers clearly for methods. Refactored class names. Added Class comments. Moved all new dialog boxes into new package to keep the gui package alittle more clean.
