+++ src/gui/dialogs/ChangePasswordDialog.java
@@ -1,6 +1,7 @@
-package gui;
+package gui.dialogs;
 
 import controller.Controller;
+import gui.AlertBox;
 import javafx.event.Event;
 import javafx.event.EventHandler;
 import javafx.geometry.Insets;
@@ -19,6 +20,8 @@ import javafx.stage.WindowEvent;
 /**
  * This is a custom dialog box that is displayed when the database has no users. This allows the new user to create
  * a new account so they can use the application. Should only appear on first run of the application.
+ *
+ * @author Prashant Bhikhu
  */
 public class ChangePasswordDialog implements EventHandler {
     // Global Components
@@ -138,7 +141,7 @@ public class ChangePasswordDialog implements EventHandler {
     }
 
 
-    String getNewPassword() {
+    public String getNewPassword() {
         return password;
     }
