+++ src/tests/parserTests.java
@@ -1,4 +1,4 @@
-package tests;
+//package tests;
 
 import static org.junit.Assert.*;
